Update sample manifest
Args instead of command, because the kube-vip binary is already executed
by the image.

Use docker.io in image path, because docker.io is not always the
default registry.

Removing Env as it is currently empty and causes the sample not to work.

Add volume mount for the config file.