Fix DHCP release bug
The macvlan interface could not be deleted until the DHCP client has
released the IP.

If the network manager in the operating system uses DHCP to obtain an IP
for the network card, it will occupy the DHCP client port. In this case
the release function will fail because the unicast connection with UDP
socket need the same port. Update dhcp package and use a raw socket.

Signed-off-by: yaocw2020 <yaocanwu@gmail.com>