Use correct address family for ipvs destinations
When IPv6 addresses are present in node.Status.Addresses, the
IPVSLoadBalancer parses the addresses and attempts to add them as an
IPv4 destinations. This results in useless entries in the service table
that only hold the last 32 bits of address.

This change does two things:
- uses the appropriate address family when creating the service
- ignores backend addresses that use a different address family than the
  service

This avoids the immediate problem of using the wrong address family for
IPv6 destinations and the useless entries in the service table.

Signed-off-by: Matthew Sykes <matthew.sykes@gmail.com>