feat: Add leaderElection retry logic
Currently if there is any failure in communications when trying to get or check the leader,
kubevip immediately fails. This causes the service to remove the vip and restart which can cause a short outage
This makes kube-vip much more reliable before failing when running in unstable or unreliable networks.

Instead of failing, if leader election fails for "retryable" conditions, then try up to 2 additional times for a total of 3

- attempt 1
- sleep 1s before attempt 2
- sleep 2s before attempt 3

if the error is something like a etcd lease update or conflict, then try much faster.
- attempt 1
- sleep 250ms before attempt 2
- sleep 500ms before attempt 3